backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
PyPDF2
pdf2image
pytesseract
scikit-learn
numpy
python-multipart
pypdf
python-dotenv
requests
groq
aiofiles